MAINT: clean up projects to start writing unit tests.
