+++ examples/CMakeLists.txt
@@ -1,4 +1,6 @@
 add_subdirectory(Features)
+add_subdirectory(FeatureDescriptors)
+add_subdirectory(FeatureDetectors)
 add_subdirectory(Geometry)
 add_subdirectory(Graphics)
 add_subdirectory(ImageProcessing)
