+++ src/DO/Sara/FeatureDetectors/Debug.hpp
@@ -14,6 +14,8 @@
 #ifndef DO_SARA_FEATUREDETECTORS_DEBUG_HPP
 #define DO_SARA_FEATUREDETECTORS_DEBUG_HPP
 
+#include <DO/Sara/Defines.hpp>
+
 
 namespace DO { namespace Sara {
 
@@ -61,20 +63,25 @@ namespace DO { namespace Sara {
   }
 
   // Check the local extrema.
+  DO_EXPORT
   void draw_scale_space_extremum(const ImagePyramid<float>& I,
                                  float x, float y, float s,
                                  int o, const Rgb8& c);
 
+  DO_EXPORT
   void draw_extrema(const ImagePyramid<float>& pyramid,
                     const std::vector<OERegion>& extrema,
                     int s, int o, bool rescaleColor = true);
 
+  DO_EXPORT
   void highlight_patch(const ImagePyramid<float>& D,
                        float x, float y, float s, int o);
 
+  DO_EXPORT
   void check_patch(const Image<float>& I, int x, int y, int w, int h,
                    double fact = 50.);
 
+  DO_EXPORT
   void check_patch(const Image<float>& I,float x, float y, float s,
                    double fact = 20.);
