+++ test/FeatureDescriptors/CMakeLists.txt
@@ -9,7 +9,7 @@ find_package(DO_Sara COMPONENTS
   Graphics ImageIO
   REQUIRED)
 
-file(GLOB test_feature_descriptors_SOURCE_FILES FILES test*.cpp)
+file(GLOB test_feature_descriptors_SOURCE_FILES FILES test_sift.cpp)
 foreach (file ${test_feature_descriptors_SOURCE_FILES})
   get_filename_component(filename "${file}" NAME_WE)
   do_add_test(${filename} ${file} "${DO_Sara_LIBRARIES}"
